feat(@angular-ru/ng-excel-builder): add column width feature
