+++ packages/ng-excel-builder/src/interfaces/column-parameters.ts
@@ -1,0 +1,5 @@
+import { ColumnWidth } from './column-width';
+
+export interface ColumnParameters {
+    width?: number | ColumnWidth;
+}
