+++ packages/ng-excel-builder/src/interfaces/column-width.ts
@@ -1,0 +1,3 @@
+export const enum ColumnWidth {
+    MAX_WIDTH = 'MAX_WIDTH'
+}
