+++ packages/ng-excel-builder/src/interfaces/excel-workbook.ts
@@ -1,14 +1,9 @@
 import { PlainObject } from '@angular-ru/common/typings';
 
-import { ExcelWorksheet, PreparedExcelWorksheet } from './excel-worksheet';
+import { ExcelWorksheet } from './excel-worksheet';
 
 export interface ExcelWorkbook<T> {
     filename: string;
     translatedKeys?: PlainObject;
     worksheets: ExcelWorksheet<T>[];
-}
-
-export interface PreparedExcelWorkbook<T> extends ExcelWorkbook<T> {
-    worksheets: PreparedExcelWorksheet<T>[];
-    preparedTranslatedKeys: PlainObject;
 }
