+++ packages/ng-excel-builder/src/interfaces/excel-worksheet.ts
@@ -1,5 +1,6 @@
-import { Nullable, PlainObject, PlainObjectOf } from '@angular-ru/common/typings';
+import { Nullable, PlainObjectOf } from '@angular-ru/common/typings';
 
+import { ColumnParameters } from './column-parameters';
 import { EntriesKeys } from './entries-keys';
 
 export interface ExcelWorksheet<T> {
@@ -9,16 +10,4 @@ export interface ExcelWorksheet<T> {
     keys?: EntriesKeys<T>;
     excludeKeys?: EntriesKeys<T>;
     columnParameters?: PlainObjectOf<ColumnParameters>;
-}
-
-export interface ColumnParameters {
-    width?: number | ColumnWidth;
-}
-
-export const enum ColumnWidth {
-    MAX_WIDTH = 'MAX_WIDTH'
-}
-
-export interface PreparedExcelWorksheet<T> extends ExcelWorksheet<T> {
-    flatEntries: PlainObject[];
 }
