+++ packages/ng-excel-builder/src/interfaces/prepared-excel-worksheet.ts
@@ -1,0 +1,7 @@
+import { PlainObject } from '@angular-ru/common/typings';
+
+import { ExcelWorksheet } from './excel-worksheet';
+
+export interface PreparedExcelWorksheet<T> extends ExcelWorksheet<T> {
+    flatEntries: PlainObject[];
+}
